feat: LEAP-1973: Adjustable spans for Text tag (#7309)
Allow users to resize selected NER regions.

The main idea is to use browser selection with styles of the region so the region expansion will look
smooth and natural. At the beginning the selection will be set to wrap the region. So users will just change
the selection during drag-n-drop, no extra code required for the visual part.

Co-authored-by: hlomzik <hlomzik@users.noreply.github.com>
Co-authored-by: bmartel <brandonmartel@gmail.com>
Co-authored-by: MihajloHoma <MihajloHoma@users.noreply.github.com>
Co-authored-by: robot-ci-heartex <robot-ci-heartex@users.noreply.github.com>