fix: LEAP-1442: Fix unexpected scrolling on region select in Video (#7392)
When user clicks on region label in Video timeline (left sidepanel of it), the video timeline scrolls to the left.
That happens because click is treated as still a click on timeline but with negative offset outside of scrollable area, so it jumps to previous window.

Also scroll to the first frame of the region on select it in region list. That worked for VideoRectangle, so here is the same code for TimelineRegion. `@todo` is the same, because it's out of the current scope.